Separate ribbon & highlight interactions. Remove unnecessary CSS code from ribbon.
